Move ffmpeg presets to utils
